requirements.txt
# Core LLM frameworks
langchain
langchain-community
langchain-chroma
langchain-pinecone
langchain-google-vertexai
google-cloud-aiplatform
transformers

# Vector stores
chromadb
pinecone-client

# Web / SSL
certifi

# Environment variables
python-dotenv

# Frontend UI
streamlit
